Add ui and document actions
