+++ client/mobile/components/document.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { View } from 'react-native';
+
+const Document = () => <View />;
+
+export default Document;
