--- client/shared/components/document.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const Document = ({ updateText }) => (
-  <div>
-    <canvas />
-    <input type="text" onKeyDown={updateText()} />
-  </div>
-);
-
-export default Document;
