+++ client/shared/reducers/errors_reducer.js
@@ -3,7 +3,7 @@ import {
   CLEAR_ERRORS
 } from '../actions/errors_actions';
 
-export default (state = [], action) => {
+const errorsReducer = (state = [], action) => {
   Object.freeze(state);
   switch (action.type) {
     case RECEIVE_ERRORS:
@@ -14,3 +14,5 @@ export default (state = [], action) => {
       return state;
   }
 };
+
+export default errorsReducer;
